Shared build service instead of cache
This increases the minimum gradle release to 6.1
