+++ gradle-conjure/src/test/groovy/com/palantir/gradle/conjure/ConjurePluginTest.groovy
@@ -572,6 +572,6 @@ class ConjurePluginTest extends IntegrationSpec {
         result.success
 
         where:
-        version << ['5.0', '5.1', '5.2', '5.3', '5.4', '5.5']
+        version << ['6.1', '6.2', '6.4', '6.5', '6.6', '6.7']
     }
 }
