+++ versions.lock
@@ -9,9 +9,8 @@ com.fasterxml.jackson.datatype:jackson-datatype-jdk8:2.11.3 (1 constraints: 7b1c
 com.fasterxml.jackson.datatype:jackson-datatype-joda:2.11.3 (1 constraints: 7b1c94a4)
 com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.11.3 (1 constraints: 7b1c94a4)
 com.fasterxml.jackson.module:jackson-module-afterburner:2.11.3 (1 constraints: 7b1c94a4)
-com.github.ben-manes.caffeine:caffeine:2.8.8 (1 constraints: 14051536)
 com.google.code.findbugs:jsr305:3.0.2 (2 constraints: 1d0fb186)
-com.google.errorprone:error_prone_annotations:2.4.0 (4 constraints: 2030ed97)
+com.google.errorprone:error_prone_annotations:2.4.0 (3 constraints: 2520701d)
 com.google.guava:failureaccess:1.0.1 (1 constraints: 140ae1b4)
 com.google.guava:guava:30.1-jre (4 constraints: 2e469540)
 com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava (1 constraints: bd17c918)
@@ -23,7 +22,7 @@ com.palantir.sls-packaging:gradle-recommended-product-dependencies:4.24.1 (1 con
 com.palantir.sls.versions:sls-versions:0.13.0 (2 constraints: b6204f28)
 joda-time:joda-time:2.9.9 (1 constraints: 07169ddb)
 net.bytebuddy:byte-buddy:1.10.18 (2 constraints: e11012e8)
-org.checkerframework:checker-qual:3.8.0 (2 constraints: 1a1a9c44)
+org.checkerframework:checker-qual:3.5.0 (1 constraints: 1a0af9b4)
 org.immutables:value:2.8.8 (1 constraints: 14051536)
 org.slf4j:slf4j-api:1.7.30 (1 constraints: f41079c3)
